Remove auto-compile feature and auto-compile-workflow.yml template (#23)
* Initial plan

* Remove auto-compile feature and auto-compile-workflow.yml template


---------
