+++ cmd/gh-aw/main.go
@@ -203,14 +203,13 @@ var compileCmd = &cobra.Command{
 		}
 		engineOverride, _ := cmd.Flags().GetString("engine")
 		validate, _ := cmd.Flags().GetBool("validate")
-		autoCompile, _ := cmd.Flags().GetBool("auto-compile")
 		watch, _ := cmd.Flags().GetBool("watch")
 		instructions, _ := cmd.Flags().GetBool("instructions")
 		if err := validateEngine(engineOverride); err != nil {
 			fmt.Fprintln(os.Stderr, console.FormatErrorMessage(err.Error()))
 			os.Exit(1)
 		}
-		if err := cli.CompileWorkflows(file, verbose, engineOverride, validate, autoCompile, watch, instructions); err != nil {
+		if err := cli.CompileWorkflows(file, verbose, engineOverride, validate, watch, instructions); err != nil {
 			fmt.Fprintln(os.Stderr, console.FormatErrorMessage(err.Error()))
 			os.Exit(1)
 		}
@@ -333,7 +332,6 @@ func init() {
 	// Add AI flag to compile and add commands
 	compileCmd.Flags().StringP("engine", "a", "", "Override AI engine (claude, codex)")
 	compileCmd.Flags().Bool("validate", false, "Enable GitHub Actions workflow schema validation")
-	compileCmd.Flags().Bool("auto-compile", false, "Generate auto-compile workflow file for automatic compilation")
 	compileCmd.Flags().BoolP("watch", "w", false, "Watch for changes to workflow files and recompile automatically")
 	compileCmd.Flags().Bool("instructions", false, "Generate or update GitHub Copilot instructions file")
