--- pkg/cli/templates/auto-compile-workflow.yml
@@ -1,35 +1,0 @@
-name: Compile Agentic Workflows
-on:
-  push:
-    branches: [main]
-    paths:
-      - '.github/workflows/*.md'
-jobs:
-  compile-agentic-workflows:
-    name: Compile Agentic Workflows
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v5
-      - name: Install gh-aw GitHub CLI extension
-        run: gh extension install githubnext/gh-aw
-      - name: Compile agentic workflows
-        run: gh aw compile --verbose
-      - name: Check for changes
-        id: changes
-        run: |
-          if git diff --quiet; then
-            echo "changed=false" >> "$GITHUB_OUTPUT"
-          else
-            echo "changed=true" >> "$GITHUB_OUTPUT"
-          fi
-      - name: Commit and push changes
-        if: steps.changes.outputs.changed == 'true'
-        run: |
-          git config --local user.email "action@github.com"
-          git config --local user.name "GitHub Action"
-          git add .github/workflows/*.lock.yml .gitattributes
-          git commit -m "Auto-compile agentic workflows [no-ci]"
-          git push
